todo/requirements.txt
asgiref==3.5.1
decouple==0.0.7
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.4
django-crispy-forms==1.14.0
django-froala-editor==4.0.11
django-on-heroku==1.1.2
django-rest-framework==0.1.0
djangorestframework==3.13.1
filelock==3.7.0
gunicorn==20.1.0
Pillow==9.1.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
virtualenv==20.14.1
whitenoise==6.1.0
